Use ManualResetEventSlim
